# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FermentOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      parameters:
        - name: style
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
...
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeWithIngredients"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeBody"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /recipes/{id}/ingredients:
    get:
      operationId: listRecipeIngredients
      tags: [ingredients]
      summary: List ingredients for a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeIngredient"
    post:
      operationId: addRecipeIngredient
      tags: [ingredients]
      summary: Add an ingredient to a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeIngredientBody"
      responses:
        "201":
          description: Ingredient added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeIngredient"
  /ingredients/{id}:
    put:
      operationId: updateRecipeIngredient
      tags: [ingredients]
      summary: Update a recipe ingredient
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeIngredientBody"
      responses:
        "200":
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeIngredient"
    delete:
      operationId: deleteRecipeIngredient
      tags: [ingredients]
      summary: Delete a recipe ingredient
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /recipes/{id}/steps:
    get:
      operationId: listRecipeSteps
      tags: [steps]
      summary: List ordered brewing steps for a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: List of steps (ordered by position)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeStep"
    post:
      operationId: addRecipeStep
      tags: [steps]
      summary: Append a brewing step to a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeStepBody"
      responses:
        "201":
          description: Step created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeStep"
  /steps/{id}:
    put:
      operationId: updateRecipeStep
      tags: [steps]
      summary: Update a brewing step
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeStepBody"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeStep"
    delete:
      operationId: deleteRecipeStep
      tags: [steps]
      summary: Delete a brewing step
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /recipes/{id}/steps/reorder:
    put:
      operationId: reorderRecipeSteps
      tags: [steps]
      summary: Replace the ordering of a recipe's steps
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRecipeStepsBody"
      responses:
        "200":
          description: Reordered steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeStep"
...
  /settings/styles:
    get:
      operationId: listBeerStyles
      tags: [settings]
      summary: List all beer styles
      responses:
        "200":
          description: List of beer styles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeerStyle"
    post:
      operationId: createBeerStyle
      tags: [settings]
      summary: Create a beer style
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeerStyleBody"
      responses:
        "201":
          description: Beer style created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeerStyle"
  /settings/styles/{id}:
    delete:
      operationId: deleteBeerStyle
      tags: [settings]
      summary: Delete a beer style
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    IngredientType:
      type: string
      enum: [malt, hop, yeast, adjunct, water_agent, other]
    IngredientUse:
      type: string
      enum: [mash, boil, dry_hop, whirlpool, primary, secondary, packaging, other]
    Recipe:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        style:
          type: string
        batchSizeGallons:
          type: number
        originalGravity:
          type: number
          nullable: true
        finalGravity:
          type: number
          nullable: true
        abv:
          type: number
          nullable: true
        ibu:
          type: number
          nullable: true
        colorSrm:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        daysPlanned:
          type: integer
          nullable: true
        daysBrewing:
          type: integer
          nullable: true
        daysFermenting:
          type: integer
          nullable: true
        daysConditioning:
          type: integer
          nullable: true
        daysPackaged:
          type: integer
          nullable: true
        avgRating:
          type: number
          nullable: true
        batchCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, style, batchSizeGallons, batchCount, createdAt, updatedAt]
    RecipeWithIngredients:
      allOf:
        - $ref: "#/components/schemas/Recipe"
        - type: object
          properties:
            ingredients:
              type: array
              items:
                $ref: "#/components/schemas/RecipeIngredient"
            steps:
              type: array
              items:
                $ref: "#/components/schemas/RecipeStep"
          required: [ingredients, steps]
    StepPhase:
      type: string
      enum: [mash, boil, fermentation, conditioning, packaging, other]
    RecipeStep:
      type: object
      properties:
        id:
          type: integer
        recipeId:
          type: integer
        position:
          type: integer
          description: 1-based ordering within the recipe
        phase:
          $ref: "#/components/schemas/StepPhase"
          nullable: true
        body:
          type: string
          description: Free-form instruction text (markdown allowed)
        durationMinutes:
          type: integer
          nullable: true
      required: [id, recipeId, position, body]
    CreateRecipeStepBody:
      type: object
      properties:
        body:
          type: string
        phase:
          $ref: "#/components/schemas/StepPhase"
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        position:
          type: integer
          nullable: true
          description: Optional explicit position; if omitted, the step is appended.
      required: [body]
    UpdateRecipeStepBody:
      type: object
      properties:
        body:
          type: string
        phase:
          $ref: "#/components/schemas/StepPhase"
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        position:
          type: integer
    ReorderRecipeStepsBody:
      type: object
      properties:
        stepIds:
          type: array
          items:
            type: integer
          description: Step IDs in the new desired order. Must contain every step ID belonging to the recipe.
      required: [stepIds]
    CreateRecipeBody:
      type: object
      properties:
        name:
          type: string
        style:
          type: string
        batchSizeGallons:
          type: number
        originalGravity:
          type: number
          nullable: true
        finalGravity:
          type: number
          nullable: true
        abv:
          type: number
          nullable: true
        ibu:
          type: number
          nullable: true
        colorSrm:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        daysPlanned:
          type: integer
          nullable: true
        daysBrewing:
          type: integer
          nullable: true
        daysFermenting:
          type: integer
          nullable: true
        daysConditioning:
          type: integer
          nullable: true
        daysPackaged:
          type: integer
          nullable: true
      required: [name, style, batchSizeGallons]
    UpdateRecipeBody:
      type: object
      properties:
        name:
          type: string
        style:
          type: string
        batchSizeGallons:
          type: number
        originalGravity:
          type: number
          nullable: true
        finalGravity:
          type: number
          nullable: true
        abv:
          type: number
          nullable: true
        ibu:
          type: number
          nullable: true
        colorSrm:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        daysPlanned:
          type: integer
          nullable: true
        daysBrewing:
          type: integer
          nullable: true
        daysFermenting:
          type: integer
          nullable: true
        daysConditioning:
          type: integer
          nullable: true
        daysPackaged:
          type: integer
          nullable: true
    RecipeIngredient:
      type: object
      properties:
        id:
          type: integer
        recipeId:
          type: integer
        name:
          type: string
        type:
          $ref: "#/components/schemas/IngredientType"
        amount:
          type: number
        unit:
          type: string
        use:
          $ref: "#/components/schemas/IngredientUse"
          nullable: true
        timingMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, recipeId, name, type, amount, unit]
    CreateRecipeIngredientBody:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/IngredientType"
        amount:
          type: number
        unit:
          type: string
        use:
          $ref: "#/components/schemas/IngredientUse"
          nullable: true
        timingMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, type, amount, unit]
    UpdateRecipeIngredientBody:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/IngredientType"
        amount:
          type: number
        unit:
          type: string
        use:
          $ref: "#/components/schemas/IngredientUse"
          nullable: true
        timingMinutes:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
...
    BeerStyle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, sortOrder, createdAt]
    CreateBeerStyleBody:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
      required: [name]
...